django-api/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==45.0.6
Django==5.2.5
django-cors-headers==4.9.0
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
dnspython==2.8.0
email-validator==2.3.0
fire==0.7.1
idna==3.10
injector==0.22.0
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.2
rav==0.0.9
rich==14.1.0
sqlparse==0.5.3
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
